Function rose::cpu::zpu::zpu_emulates

pub fn zpu_emulates(zpu: &mut ZPU, op: u8) -> Result<bool, Error>

Emulates.

ZPU EMULATE operations are optional for the most part, but hardware implementations are obviously quicker than software ones. Apart from that, not all software implements those, relying on hardware implementations instead.